requirements.txt
bokeh>=2.0.1
html5lib>=1.0.1
importlib-metadata>=1.6.0
jupyter>=1.0.0
Keras>=2.2.4
matplotlib>=3.2.1
numpy>=1.18.2
pandas>=1.0.3
plaidml>=0.7.0
plaidml-keras>=0.7.0
plumbum>=1.6.9
prometheus-client>=0.7.1
pytest>=5.4.2
python-dateutil>=2.8.1
PyYAML>=5.3.1
scikit-learn>=0.22.2.post1
scipy>=1.4.1
seaborn>=0.10.0
selenium>=3.141.0
six>=1.14.0
urllib3>=1.25.8
requests-mock>=1.8.0
